refactor: add layout file to setup register layout
